Wed, Sep 28, 2022, 12:57 PM +00:00

+++ sentimental-cash/cash.py
@@ -6,7 +6,7 @@ def main():
 
 def get_cents():
     while True:
-        n = int(input("No. of Cents:"))
+        n = float(input("No. of Cents:"))
         if n > 0:
              break
